# AdminNotificator/AdminNotificator.WebApi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
    "/notifications": {
      "get": {
        "tags": [
          "Email"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmailType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "name": "ExperianceDays",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "EmailTitle",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IntersectDepartmentIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "ExceptDepartmentIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "IntersectOrganizationNames",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "BodyName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IntersectTowns",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "ExceptTowns",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "DontSendAfterDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "MaternityDays",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ForGenders",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "DayCountsForResend",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "IntersectUserPosts",
            "in": "query",
            "schema": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "ExceptUserPosts",
            "in": "query",
            "schema": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "SenderEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Bcc",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailType"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/{id}": {
      "get": {
        "tags": [
          "Email"
        ],
        "operationId": "GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailType"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{id}": {
      "put": {
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ExperianceDays",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "EmailTitle",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IntersectDepartmentIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "ExceptDepartmentIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "IntersectOrganizationNames",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "BodyName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IntersectTowns",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "ExceptTowns",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "DontSendAfterDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "MaternityDays",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ForGenders",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "DayCountsForResend",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "IntersectUserPosts",
            "in": "query",
            "schema": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "ExceptUserPosts",
            "in": "query",
            "schema": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "SenderEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Bcc",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailType"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/deleteMany": {
      "post": {
        "tags": [
          "Email"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmailType": {
        "required": [
          "bodyName",
          "emailTitle",
          "id",
          "senderEmail"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "experianceDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "emailTitle": {
            "type": "string",
            "nullable": true
          },
          "intersectDepartmentIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "exceptDepartmentIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "intersectOrganizationNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "bodyName": {
            "type": "string",
            "nullable": true
          },
          "intersectTowns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "exceptTowns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "dontSendAfterDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "maternityDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "forGenders": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "dayCountsForResend": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "intersectUserPosts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "exceptUserPosts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "senderEmail": {
            "type": "string",
            "nullable": true
          },
          "bcc": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}